refactor(mseed.py): rename module
Rename io/mseed.py to io/data.py. This better reflects what the module
is used for, as in theory any waveform data format can be used.
